+++ components/ui/NavComponent/NavbarLogin.tsx
@@ -1,5 +1,4 @@
-import * as React from "react";
-import { useContext } from "react";
+import { useRef, useState, useContext } from "react";
 import { AvatarImage, Avatar } from "../avatar";
 import { Popover, PopoverContent, PopoverTrigger } from "../popover";
 import LoginContext from "@/utils/contexts/login";
@@ -7,13 +6,22 @@ import { Button } from "@/components/ui/button";
 import { QuickLoginDialog } from "@/components/ui/quicklogindialog";
 import { capitalizeFirstLetter } from "@/utils/utils";
 import { NAV_ELEMENTS_VARIANT } from "@/utils/constants";
-import { LoginComponent } from "@/components/ui/logincomponent";
-
 import { useIsMobile } from "@/components/hooks/use-mobile";
 import { Sheet, SheetContent, SheetClose, SheetTrigger } from "@/components/ui/sheet";
+import { Input } from "@/components/ui/input";
+import { COMPANY_LOGOS } from "@/utils/constants";
 
 const NavBarLoginInterface = () => {
-    const { isLoggedIn, userObject, logoutUser } = useContext(LoginContext);
+    const { isLoggedIn, userObject, logoutUser, loginUser } = useContext(LoginContext);
+
+    const inputRef = useRef<HTMLInputElement>(null);
+    const [defaultEmail, setDefaultEmail] = useState<string>("user@launchmail.io");
+
+    function handleLogin(): void {
+        if (!defaultEmail) return;
+
+        loginUser("user@launchmail.io");
+    }
 
     return (
         <>
@@ -47,7 +55,45 @@ const NavBarLoginInterface = () => {
                 </>
             )}
 
-            {!isLoggedIn && <LoginComponent />}
+            {!isLoggedIn && (
+                <div className="w-full  bg-white font-audimat shadow-2xl mx-auto text-black p-4 sm:p-8 h-full flex flex-col">
+                    <div className=" mx-auto text-center mt-4 mb-8">
+                        <img src={COMPANY_LOGOS["bank"].vertical.src} className=" mx-auto" />
+                    </div>
+                    <div className="w-full flex flex-col gap-y-4 mb-4">
+                        <Input
+                            placeholder="Email"
+                            value={defaultEmail}
+                            ref={inputRef}
+                            required
+                            className=" outline-none border-0 border-b-2 text-xl"
+                            onChange={(e) => setDefaultEmail(e.target.value)}
+                        />
+
+                        <Button
+                            onClick={() => handleLogin()}
+                            className={` w-full mx-auto font-sohnelight rounded-none  text-lg bg-loginComponentBlue text-white`}
+                        >
+                            Login with SSO
+                        </Button>
+                        <QuickLoginDialog />
+                    </div>
+                    <div className="flex flex-row items-start sm:items-baseline
+                     font-sohnelight font-extralight sm:flex-col text-xs justify-between gap-y-2">
+                        <div className="">
+                            <p>Forgot Password?</p>
+                        </div>
+                        <div>
+                            <p className="text-right flex flex-col sm:flex-row ">
+                                Don't have an account?{" "}
+                                <a href="#" className=" ml-2 cursor-auto">
+                                    Sign Up
+                                </a>
+                            </p>
+                        </div>
+                    </div>
+                </div>
+            )}
         </>
     );
 };
